refactor: Split GraphQL schema and queries into modular files
- Split lib/graphql/queries.ts (1019 lines) into modular files:
  - queries/fragments.ts - GraphQL fragments
  - queries/person.ts - Person queries/mutations
  - queries/family.ts - Family queries/mutations
  - queries/source.ts - Source mutations
  - queries/tree.ts - Tree queries
  - queries/dashboard.ts - Dashboard/stats queries
  - queries/admin.ts - Admin queries/mutations
  - queries/crest.ts - Coat of arms queries/mutations
  - queries/settings.ts - Settings and GEDCOM
  - queries/media.ts - Media mutations
  - queries/index.ts - Re-exports for backward compatibility

- Split lib/graphql/schema.ts (781 lines) into modular files:
  - schema/types/person.ts - Person type definitions
  - schema/types/family.ts - Family type definitions
  - schema/types/source.ts - Source type definitions
  - schema/types/media.ts - Media type definitions
  - schema/types/comment.ts - Comment type definitions
  - schema/types/crest.ts - Crest type definitions
  - schema/types/admin.ts - Admin/user type definitions
  - schema/types/settings.ts - Settings type definitions
  - schema/types/common.ts - Pagination, stats, dashboard, tree types
  - schema/queries.ts - Query type definitions
  - schema/mutations.ts - Mutation type definitions
  - schema/index.ts - Combines all types

- Updated test to use new modular schema path
- Maintains full backward compatibility via re-exports

Closes #303
